@@ -15,53 +15,58 @@
   ],
   "scripts": {
     "test": "jest",
-    "build": "gulp && gulp",
-    "package": "npm run build && gulp package",
-    "release": "git tag -a v10.0.1 -m 'Releasing version v10.0.1' && git push origin --tags && git push origin master"
+    "build": "webpack"
   },
   "devDependencies": {
-    "@babel/plugin-transform-runtime": "^7.2.0",
-    "babel-core": "^6.0.0",
-    "babel-loader": "^6.4.1",
+    "@babel/core": "^7.11.6",
+    "@babel/plugin-proposal-optional-chaining": "^7.11.0",
+    "@babel/plugin-transform-runtime": "^7.11.0",
+    "@babel/preset-env": "^7.4.5",
+    "@babel/register": "^7.4.4",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-jest": "^24.8.0",
+    "babel-loader": "^8.1.0",
+    "babel-minify-webpack-plugin": "^0.3.1",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.7.0",
-    "babel-preset-es2015": "^6.6.0",
-    "cross-env": "^3.0.0",
-    "css-loader": "^0.25.0",
-    "del": "^2.2.0",
-    "extract-text-webpack-plugin": "*",
-    "file-loader": "^0.9.0",
-    "gulp": "^3.9.1",
-    "gulp-babel": "^6.1.2",
-    "gulp-cache": "^0.4.3",
-    "gulp-chrome-manifest": "0.0.13",
-    "gulp-clean-css": "^2.0.3",
-    "gulp-eslint": "^2.0.0",
-    "gulp-htmlmin": "^1.3.0",
-    "gulp-if": "^2.0.0",
-    "gulp-imagemin": "^2.4.0",
-    "gulp-livereload": "^3.8.1",
-    "gulp-load-plugins": "^1.2.0",
-    "gulp-sass": "^3.1.0",
-    "gulp-size": "^2.1.0",
-    "gulp-sourcemaps": "^1.6.0",
-    "gulp-uglify": "^1.5.3",
-    "gulp-useref": "^3.0.8",
-    "gulp-zip": "^3.2.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "core-js": "^2.6.5",
+    "cross-env": "^5.2.0",
+    "css-loader": "^3.0.0",
+    "deepmerge": "^4.2.2",
+    "del": "^4.1.1",
+    "eslint": "^7.7.0",
+    "eslint-config-prettier": "^5.0.0",
+    "eslint-plugin-prettier": "^3.1.1",
+    "eslint-plugin-vue": "^7.0.0-beta.2",
+    "eslint-plugin-vue-a11y": "0.0.31",
+    "eslint-plugin-vuejs-accessibility": "^0.4.0",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
+    "fibers": "^4.0.2",
+    "file-loader": "^4.0.0",
+    "jest": "^24.8.0",
+    "main-bower-files": "^2.13.1",
+    "mark.js": "^8.11.1",
     "purify-css": "^1.2.5",
     "purifycss-webpack": "^0.7.0",
-    "run-sequence": "^1.1.5",
-    "sass-loader": "^7.1.0",
-    "style-loader": "^0.21.0",
-    "uglifyjs-webpack-plugin": "^1.0.0",
-    "vue": "^2.5.16",
-    "vue-lazyload": "^1.2.6",
-    "vue-loader": "^12.2.2",
-    "vue-template-compiler": "^2.5.16",
-    "webpack": "^3.4.1",
-    "webpack-bundle-analyzer": "^3.0.3",
-    "webpack-dev-server": "^2.4.5",
+    "run-sequence": "^2.2.1",
+    "sass": "^1.23.3",
+    "sass-loader": "^8.0.0",
+    "style-loader": "^0.23.1",
+    "tesseract.js": "^2.0.0-beta.1",
+    "vue-jest": "^3.0.4",
+    "vue-loader": "^15.7.0",
+    "vue-template-compiler": "^2.6.10",
+    "webpack": "^4.44.1",
+    "webpack-bundle-analyzer": "^3.8.0",
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.7.1",
+    "webpack-zip-plugin": "^0.1.6",
     "wiredep": "^4.0.0",
-    "xo": "^0.18.2"
+    "xo": "^0.24.0",
+    "zip-webpack-plugin": "^3.0.0"
   },
   "eslintConfig": {
     "env": {
@@ -80,26 +85,68 @@
     }
   },
   "dependencies": {
+    "@babel/runtime": "^7.5.5",
     "@fortawesome/fontawesome-svg-core": "^1.2.26",
     "@fortawesome/free-brands-svg-icons": "^5.12.0",
     "@fortawesome/free-solid-svg-icons": "^5.12.0",
     "@fortawesome/vue-fontawesome": "^0.1.9",
+    "@microsoft/immersive-reader-sdk": "^1.0.0",
+    "@stripe/stripe-js": "^1.7.0",
+    "async": "^3.1.0",
     "axios": "^0.19.0",
     "beautify-scrollbar": "^1.0.8",
     "bootstrap-vue": "^2.1.0",
+    "cheet.js": "^0.3.3",
+    "copy-webpack-plugin": "^5.1.1",
     "font-awesome": "^4.7.0",
+    "fs": "0.0.1-security",
+    "github-pages-deploy-action": "^3.4.5",
+    "google-translate": "^3.0.0",
+    "gulp-concat": "^2.6.1",
+    "gulp-replace": "^1.0.0",
+    "hash-files": "^1.1.1",
+    "husky": "^3.1.0",
     "i": "^0.3.6",
-    "material-icons": "^0.3.0",
-    "npm": "^6.4.1",
-    "v2-lazy-list": "^2.2.5",
-    "vee-validate": "^3.2.0",
-    "vue-loading-spinner": "^1.0.11",
-    "vue-material": "^1.0.0-beta-10.2",
+    "install": "^0.13.0",
+    "jcrop-0.9.12": "^0.9.12",
+    "jquery": "^3.1.1",
+    "jsdom": "^15.2.1",
+    "lodash.clonedeep": "^4.5.0",
+    "marked": "^0.6.2",
+    "material-design-icons": "^3.0.1",
+    "material-icons": "^0.3.1",
+    "moment": "^2.24.0",
+    "moment-mini": "^2.24.0",
+    "notyf": "^3.6.0",
+    "npm": "^6.9.0",
+    "prettier": "^1.19.1",
+    "pretty-quick": "^2.0.1",
+    "read-time-estimate": "0.0.2",
+    "readability": "github:mozilla/readability",
+    "rome": "^10.0.4-beta",
+    "sinon-chrome": "^3.0.1",
+    "stripe": "^7.15.0",
+    "tesseract.js": "^1.0.14",
+    "traverse": "^0.6.6",
+    "uncss-webpack-plugin": "^0.2.0",
+    "verte": "0.0.12",
+    "vue": "^2.6.10",
+    "vue-color": "^2.7.1",
+    "vue-html-editor": "^0.2.1",
+    "vue-loaders": "^3.0.2",
+    "vue-material": "^1.0.0-beta-15",
+    "vue-material-slider": "^1.0.1",
     "vue-plugin-webextension-i18n": "^0.1.2",
-    "vue-resource": "^1.5.1",
-    "vue-ripple-directive": "^2.0.1",
-    "vue-select": "^3.10.3",
+    "vue-print-nb": "^1.3.0",
+    "vue-select": "^3.10.8",
+    "vue-slider-component": "^3.2.5",
     "vue-spinner": "^1.0.3",
-    "webpack-stream": "^5.1.1"
+    "vue-stripe-checkout": "^1.2.6",
+    "vue-swatches": "^1.0.4",
+    "vue2-editor": "^2.10.2",
+    "vuelidate": "^0.7.5",
+    "vuetify": "^1.5.20",
+    "webpack-babel-jest": "^1.0.4",
+    "webpack-stream": "^5.2.1"
   }
 }
